refactor: baseballGameCore와 baseballGame 합치는 작업 수행
